# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления задачами и тегами",
        "title": "Task Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@taskmanager.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks": {
            "get": {
                "description": "Получение списка задач с возможностью фильтрации и пагинации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Список задач с фильтрацией",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Статус задачи (new, in_progress, done, cancelled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата начала периода (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата конца периода (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Теги для фильтрации",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Лимит записей",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение для пагинации",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.TaskListResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создание новой задачи",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Создать задачу",
                "parameters": [
                    {
                        "description": "Данные задачи",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.Task"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks-period": {
            "get": {
                "description": "Получение списка задач за указанный период",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks-period"
                ],
                "summary": "Получить задачи за период",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Начало периода (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Конец периода (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.TaskListResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создание задачи с указанием периода выполнения",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks-period"
                ],
                "summary": "Создать задачу с периодом",
                "parameters": [
                    {
                        "description": "Данные задачи с периодом",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.CreateTaskPeriodRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.Task"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks-period/{id}": {
            "put": {
                "description": "Установка статуса задачи на конкретную дату",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks-period"
                ],
                "summary": "Установить статус задачи на дату",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Статус задачи",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.SetStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаление переопределения статуса задачи на конкретную дату",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks-period"
                ],
                "summary": "Удалить переопределение статуса задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата переопределения (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Not found"
                }
            }
        },
...
        "github_com_Vladroon22_TaskTracker_internal_models.CreateTaskPeriodRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "is_recurring": {
                    "type": "boolean"
                },
                "recurrence_days": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "recurrence_interval": {
                    "type": "integer"
                },
                "recurrence_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_Vladroon22_TaskTracker_internal_models.CreateTaskRequest": {
            "type": "object",
            "required": [
                "due_date",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "is_recurring": {
                    "type": "boolean"
                },
                "recurrence_days": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "recurrence_interval": {
                    "type": "integer"
                },
                "recurrence_type": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.TaskStatus"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid request body"
                }
            }
        },
...
        "github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
        "github_com_Vladroon22_TaskTracker_internal_models.SetStatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "done"
                }
            }
        },
...
        "github_com_Vladroon22_TaskTracker_internal_models.Task": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_recurring": {
                    "type": "boolean"
                },
                "recurrence_days": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "recurrence_interval": {
                    "type": "integer"
                },
                "recurrence_type": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.TaskStatus"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.Tag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_Vladroon22_TaskTracker_internal_models.TaskListResponse": {
            "type": "object",
            "properties": {
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.Task"
                    }
                }
            }
        },
...
    },
...
}